Extract marketing subscriptions into ATTNSDK+MarketingSubscriptions.swift
Fixes SwiftLint type_body_length error (707 lines, limit 700) caused by
MSDK-344 pushEnabled additions combined with MSDK-315 error handling.
